fixup! .NET C# stacktrace support #5489
