+++ ide/che-core-ide-app/src/main/java/org/eclipse/che/ide/console/AbstractOutputCustomizer.java
@@ -1,18 +1,17 @@
 /*******************************************************************************
- * Copyright (c) 2017 RedHat, Inc.
+ * Copyright (c) 2017 Red Hat, Inc.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  *
  * Contributors:
- *   RedHat, Inc. - initial API and implementation
+ *   Red Hat, Inc. - initial API and implementation
  *******************************************************************************/
 package org.eclipse.che.ide.console;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Strings.nullToEmpty;
-import static com.google.gwt.regexp.shared.RegExp.compile;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -34,10 +33,7 @@ import org.eclipse.che.ide.api.resources.File;
 import org.eclipse.che.ide.api.resources.Resource;
 import org.eclipse.che.ide.resource.Path;
 
-import com.google.gwt.regexp.shared.MatchResult;
-import com.google.gwt.regexp.shared.RegExp;
 import com.google.gwt.user.client.Timer;
-import com.google.inject.Inject;
 
 /**
  * Default customizer adds an anchor link to the lines that match a stack trace
@@ -45,20 +41,17 @@ import com.google.inject.Inject;
  * the stack trace line, searches for the candidate Java files to navigate to,
  * opens the first file (of the found candidates) in editor and reveals it to
  * the required line according to the stack trace line information
+ * 
+ * @author Victor Rubezhny
  */
-public class DefaultOutputCustomizer implements OutputCustomizer {
+abstract public class AbstractOutputCustomizer implements OutputCustomizer {
 
-    private static final RegExp LINE_AT = compile("(\\s+at .+\\(.+\\.java:\\d+\\))");
+    protected AppContext appContext;
+    protected EditorAgent editorAgent;
 
-    private AppContext appContext;
-    private EditorAgent editorAgent;
-
-    @Inject
-    public DefaultOutputCustomizer(AppContext appContext, EditorAgent editorAgent) {
+    public AbstractOutputCustomizer(AppContext appContext, EditorAgent editorAgent) {
         this.appContext = appContext;
         this.editorAgent = editorAgent;
-
-        exportAnchorClickHandlerFunction();
     }
 
     /*
@@ -68,9 +61,7 @@ public class DefaultOutputCustomizer implements OutputCustomizer {
      * OutputCustomizer#canCustomize(java.lang.String)
      */
     @Override
-    public boolean canCustomize(String text) {
-        return (LINE_AT.exec(text) != null);
-    }
+    abstract public boolean canCustomize(String text);
 
     /*
      * (non-Javadoc)
@@ -79,66 +70,12 @@ public class DefaultOutputCustomizer implements OutputCustomizer {
      * OutputCustomizer#customize(java.lang.String)
      */
     @Override
-    public String customize(String text) {
-        MatchResult matcher = LINE_AT.exec(text);
-        if (matcher != null) {
-            try {
-                int start = text.indexOf("at", 0) + "at".length(), openBracket = text.indexOf("(", start),
-                        column = text.indexOf(":", openBracket), closingBracket = text.indexOf(")", column);
-                String qualifiedName = text.substring(start, openBracket).trim();
-                String fileName = text.substring(openBracket + "(".length(), column).trim();
-                int lineNumber = Integer.valueOf(text.substring(column + ":".length(), closingBracket).trim());
-
-                String customText = text.substring(0, openBracket + "(".length());
-                customText += "<a href='javascript:open(\"" + qualifiedName + "\", \"" + fileName + "\", " + lineNumber
-                        + ");'>";
-                customText += text.substring(openBracket + "(".length(), closingBracket);
-                customText += "</a>";
-                customText += text.substring(closingBracket);
-                text = customText;
-            } catch (IndexOutOfBoundsException ex) {
-                // ignore
-            } catch (NumberFormatException ex) {
-                // ignore
-            }
-        }
-
-        return text;
-    }
-
-    /**
-     * A callback that is to be called for an anchor
-     * 
-     * @param qualifiedName
-     * @param fileName
-     * @param lineNumber
-     */
-    public void handleAnchorClick(String qualifiedName, String fileName, final int lineNumber) {
-        if (qualifiedName == null || fileName == null) {
-            return;
-        }
-
-        String qualifiedClassName = qualifiedName.lastIndexOf('.') != -1
-                ? qualifiedName.substring(0, qualifiedName.lastIndexOf('.'))
-                : qualifiedName;
-        final String packageName = qualifiedClassName.lastIndexOf('.') != -1
-                ? qualifiedClassName.substring(0, qualifiedClassName.lastIndexOf('.'))
-                : "";
-
-        String relativeFilePath = (packageName.isEmpty() ? "" : (packageName.replace(".", "/") + "/")) + fileName;
-
-        collectChildren(appContext.getWorkspaceRoot(), Path.valueOf(relativeFilePath))
-        .then(files -> {
-            if (!files.isEmpty()) {
-                openFileInEditorAndReveal(appContext, editorAgent, files.get(0).getLocation(), lineNumber, 0);
-            }
-        });
-    }
+    abstract public String customize(String text);
 
     /*
      * Returns the list of workspace files filtered by a relative path
      */
-    private Promise<List<File>> collectChildren(Container root, Path relativeFilePath) {
+    protected Promise<List<File>> collectChildren(Container root, Path relativeFilePath) {
         return root.getTree(-1).then(new Function<Resource[], List<File>>() {
             @Override
             public List<File> apply(Resource[] children) throws FunctionException {
@@ -152,7 +89,7 @@ public class DefaultOutputCustomizer implements OutputCustomizer {
     /*
      * Checks if a path's last segments are equal to the provided relative path
      */
-    private boolean endsWith(Path path, Path relativePath) {
+    protected boolean endsWith(Path path, Path relativePath) {
         checkNotNull(path);
         checkNotNull(relativePath);
 
@@ -176,7 +113,7 @@ public class DefaultOutputCustomizer implements OutputCustomizer {
      * @param lineNumber
      * @param columnNumber
      */
-    static void openFileInEditorAndReveal(AppContext appContext, EditorAgent editorAgent, Path file,
+    protected void openFileInEditorAndReveal(AppContext appContext, EditorAgent editorAgent, Path file,
             final int lineNumber, final int columnNumber) {
         appContext.getWorkspaceRoot().getFile(file).then(optional -> {
             if (optional.isPresent()) {
@@ -209,7 +146,7 @@ public class DefaultOutputCustomizer implements OutputCustomizer {
      * @param line
      * @param column
      */
-    static void selectRange(EditorPartPresenter editor, int line, int column) {
+    protected void selectRange(EditorPartPresenter editor, int line, int column) {
         line--;
         column--;
         if (line < 0)
@@ -224,27 +161,4 @@ public class DefaultOutputCustomizer implements OutputCustomizer {
             document.setCursorPosition(new TextPosition(line, column >= 0 ? column : 0));
         }
     }
-
-    /*
-     * Selects and shows the specified line and column of text in editor
-     */
-    private void selectRange(EditorPartPresenter editor, int line) {
-        line--;
-        if (editor instanceof TextEditor) {
-            Document document = ((TextEditor) editor).getDocument();
-            LinearRange selectionRange = document.getLinearRangeForLine(line);
-            document.setCursorPosition(new TextPosition(line, 0));
-            document.setSelectedRange(selectionRange, true);
-        }
-    }
-
-    /**
-     * Sets up a java callback to be called for an anchor
-     */
-    public native void exportAnchorClickHandlerFunction() /*-{
-        var that = this;
-        $wnd.open = $entry(function(qualifiedName,fileName,lineNumber) {
-            that.@org.eclipse.che.ide.console.DefaultOutputCustomizer::handleAnchorClick(*)(qualifiedName,fileName,lineNumber);
-        });
-    }-*/;
 }
