+++ ide/che-core-ide-app/src/main/java/org/eclipse/che/ide/console/OutputCustomizer.java
@@ -1,17 +1,19 @@
 /*******************************************************************************
- * Copyright (c) 2017 RedHat, Inc.
+ * Copyright (c) 2017 Red Hat, Inc.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  *
  * Contributors:
- *   RedHat, Inc. - initial API and implementation
+ *   Red Hat, Inc. - initial API and implementation
  *******************************************************************************/
 package org.eclipse.che.ide.console;
 
 /**
  * An interface to allow output text customizations 
+ * 
+ * @author Victor Rubezhny
  */
 public interface OutputCustomizer {
     /** Checks if the specified text can be/has to be customized */
