+++ ide/che-core-ide-app/src/test/java/org/eclipse/che/ide/console/CSharpOutputCustomizerTest.java
@@ -1,12 +1,12 @@
 /*******************************************************************************
- * Copyright (c) 2017 RedHat, Inc.
+ * Copyright (c) 2017 Red Hat, Inc.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  *
  * Contributors:
- *   RedHat, Inc. - initial API and implementation
+ *   Red Hat, Inc. - initial API and implementation
  *******************************************************************************/
 package org.eclipse.che.ide.console;
 
@@ -24,19 +24,20 @@ import com.google.gwtmockito.GwtMockitoTestRunner;
  * CSharpOutputCustomazer.
  * 
  * See: Issue #5489 - .NET C# stacktrace support #5489
+ * 
+ * @author Victor Rubezhny
  */
 @RunWith(GwtMockitoTestRunner.class)
-public class CSharpOutputCustomizerTest {
+public class CSharpOutputCustomizerTest extends BaseOutputCustomizerTest {
     @Mock
     AppContext appContext;
     @Mock
     EditorAgent editorAgent;
 
-    private OutputCustomizer outputCustomizer;
-
     @Before
     public void setUp() throws Exception {
-        outputCustomizer = new CSharpOutputCustomizer(appContext, editorAgent);
+        OutputCustomizer outputCustomizer = new CSharpOutputCustomizer(appContext, editorAgent);
+        setupTestCustomizers(outputCustomizer, new OutputCustomizer[] { outputCustomizer });
     }
 
     /**
@@ -49,9 +50,8 @@ public class CSharpOutputCustomizerTest {
      */
     @Test
     public void testInitialStackTraceLines() throws Exception {
-        DefaultOutputCustomizerTest.testStackTraceLine(outputCustomizer, false,
-                "Unhandled Exception: System.NullReferenceException: Object reference not set to an instance of an object.",
-                null);
+        testStackTraceLine(
+                "Unhandled Exception: System.NullReferenceException: Object reference not set to an instance of an object.");
     }
 
     /**
@@ -63,10 +63,10 @@ public class CSharpOutputCustomizerTest {
      */
     @Test
     public void testValuableStackTraceLines() throws Exception {
-        DefaultOutputCustomizerTest.testStackTraceLine(outputCustomizer, true,
+        testStackTraceLine(CSharpOutputCustomizer.class,
                 "   at hwapp.ppp.PPPProgram.Main1(String[] args) in /home/jeremy/projects/csharp/hwapp/ppp/PPPProgram.cs:line 18",
                 "   at hwapp.ppp.PPPProgram.Main1(String[] args) in <a href='javascript:openCSSTL(\"/home/jeremy/projects/csharp/hwapp/ppp/PPPProgram.cs\",18);'>/home/jeremy/projects/csharp/hwapp/ppp/PPPProgram.cs:line 18</a>");
-        DefaultOutputCustomizerTest.testStackTraceLine(outputCustomizer, true,
+        testStackTraceLine(CSharpOutputCustomizer.class,
                 "   at hwapp.Program.Main(String[] args) in /home/jeremy/projects/csharp/hwapp/Program.cs:line 10",
                 "   at hwapp.Program.Main(String[] args) in <a href='javascript:openCSSTL(\"/home/jeremy/projects/csharp/hwapp/Program.cs\",10);'>/home/jeremy/projects/csharp/hwapp/Program.cs:line 10</a>");
     }
@@ -81,10 +81,10 @@ public class CSharpOutputCustomizerTest {
      */
     @Test
     public void testValuableCompilationMessagesLines() throws Exception {
-        DefaultOutputCustomizerTest.testStackTraceLine(outputCustomizer, true,
+        testStackTraceLine(CSharpOutputCustomizer.class,
                 "Program.cs(2,13): error CS0234: The type or namespace name 'ppp' does not exist in the namespace 'hwapp' (are you missing an assembly reference?) [/home/jeremy/projects/csharp/hwapp/hwapp.csproj]",
                 "<a href='javascript:openCSCM(\"Program.cs\",\"/home/jeremy/projects/csharp/hwapp/hwapp.csproj\",2,13);'>Program.cs(2,13)</a>: error CS0234: The type or namespace name 'ppp' does not exist in the namespace 'hwapp' (are you missing an assembly reference?) [/home/jeremy/projects/csharp/hwapp/hwapp.csproj]");
-        DefaultOutputCustomizerTest.testStackTraceLine(outputCustomizer, true,
+        testStackTraceLine(CSharpOutputCustomizer.class,
                 "ppp/PPPProgram.cs(9,17): warning CS0219: The variable 'testIntValue' is assigned but its value is never used [/home/jeremy/projects/csharp/hwapp/hwapp.csproj]",
                 "<a href='javascript:openCSCM(\"ppp/PPPProgram.cs\",\"/home/jeremy/projects/csharp/hwapp/hwapp.csproj\",9,17);'>ppp/PPPProgram.cs(9,17)</a>: warning CS0219: The variable 'testIntValue' is assigned but its value is never used [/home/jeremy/projects/csharp/hwapp/hwapp.csproj]");
     }
@@ -98,22 +98,22 @@ public class CSharpOutputCustomizerTest {
      */
     @Test
     public void testNonCSharpStackTraceLines() throws Exception {
-        DefaultOutputCustomizerTest.testStackTraceLine(outputCustomizer, false,
-                "[STDOUT] Listening for transport dt_socket at address: 4403", null);
-        DefaultOutputCustomizerTest.testStackTraceLine(outputCustomizer, false,
-                "org.test.HighLevelException: org.test.MidLevelException: org.test.LowLevelException", null);
-        DefaultOutputCustomizerTest.testStackTraceLine(outputCustomizer, false,
-                "Caused by: org.test.MidLevelException: org.test.LowLevelException", null);
-        DefaultOutputCustomizerTest.testStackTraceLine(outputCustomizer, false, 
-                "Caused by: org.test.LowLevelException", null);
-        DefaultOutputCustomizerTest.testStackTraceLine(outputCustomizer, false,
-                "Exception in thread \"main\" java.lang.ArithmeticException: / by zero", null);
-        DefaultOutputCustomizerTest.testStackTraceLine(outputCustomizer, false, 
-                "   at org.test.Junk.main(Junk.java:6)", null);
-        DefaultOutputCustomizerTest.testStackTraceLine(outputCustomizer, false,
-                "   at org.test.TrashClass.throwItThere(Junk.java:51)", null);
-        DefaultOutputCustomizerTest.testStackTraceLine(outputCustomizer, false,
-                "   at MyClass$ThrowInConstructor.<init>(MyClass.java:16)", null);
-        DefaultOutputCustomizerTest.testStackTraceLine(outputCustomizer, false, "   ... 1 more", null);
+        testStackTraceLine(
+                "[STDOUT] Listening for transport dt_socket at address: 4403");
+        testStackTraceLine(
+                "org.test.HighLevelException: org.test.MidLevelException: org.test.LowLevelException");
+        testStackTraceLine(
+                "Caused by: org.test.MidLevelException: org.test.LowLevelException");
+        testStackTraceLine( 
+                "Caused by: org.test.LowLevelException");
+        testStackTraceLine(
+                "Exception in thread \"main\" java.lang.ArithmeticException: / by zero");
+        testStackTraceLine( 
+                "   at org.test.Junk.main(Junk.java:6)");
+        testStackTraceLine(
+                "   at org.test.TrashClass.throwItThere(Junk.java:51)");
+        testStackTraceLine(
+                "   at MyClass$ThrowInConstructor.<init>(MyClass.java:16)");
+        testStackTraceLine("   ... 1 more");
     }
 }
